refactor(expressions): Use kork-expressions
