+++ orca-core/orca-core.gradle
@@ -33,6 +33,7 @@ dependencies {
   compile spinnaker.dependency('spectatorApi')
   compile spinnaker.dependency('kork')
   compile spinnaker.dependency('korkExceptions')
+  compile spinnaker.dependency('korkExpressions')
   compile spinnaker.dependency('korkArtifacts')
   compile spinnaker.dependency("logstashEncoder")
   compile spinnaker.dependency("springSecurityWeb")
