--- orca-core/src/main/java/com/netflix/spinnaker/orca/pipeline/util/ContextFunctionConfiguration.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2018 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.orca.pipeline.util;
-
-import com.netflix.spinnaker.orca.config.UserConfiguredUrlRestrictions;
-import static com.netflix.spinnaker.orca.pipeline.expressions.PipelineExpressionEvaluator.ExpressionEvaluationVersion.V2;
-
-public class ContextFunctionConfiguration {
-  private final UserConfiguredUrlRestrictions urlRestrictions;
-  private final String spelEvaluator;
-
-  public ContextFunctionConfiguration(UserConfiguredUrlRestrictions urlRestrictions, String spelEvaluator) {
-    this.urlRestrictions = urlRestrictions;
-    this.spelEvaluator = spelEvaluator;
-  }
-
-  public ContextFunctionConfiguration(UserConfiguredUrlRestrictions urlRestrictions) {
-    this(urlRestrictions, V2);
-  }
-
-  public UserConfiguredUrlRestrictions getUrlRestrictions() {
-    return urlRestrictions;
-  }
-
-  public String getSpelEvaluator() {
-    return spelEvaluator;
-  }
-}
